tested and fixed running test_tflite_detect.py
uploaded and installed basic_bot on raspberry pi4 raspian bullseye 64bit.  It works and with the tweaks made should be able to run with either version of tensor flow lite that runs in the the unbutu test container on ci/cd and on the older version that runs on raspian bullseye.

still no joy running tensor-flow lite on apple silicon tho :(