Picamera2 stack working 🎉
Tested on pi5 Bookworm (64bit). The service test is unfortunately not able to run on mac, but I was able to run all of our tests on a pi5 includng the vision service tests.  The stats on camera capture using the bookworm->picamera2->tflight are looking really great - 28fps on capture and 26fps on recognition. 🍾