Hub should stop sending to star subscribers
After a star subscriber disconnects or is disconnected, central_hub should stop trying to relay state updates.  I caught this bug while testing D2 and noticed in the central_hub log that it was still trying to relay state updates to a closed browser client.  I'm not sure how to test for it from a harness point of view except maybe client-1 connect+subscribe_all+disconnect and client-2 connect+send_update+?  maybe scan the log for an error?  client-1 will still not recieve anything if it reconnects, but the hub will still try.  I confirmed via manual test that this PR fixes the issue.