maybe fix tests in CI/CD -\0/- ...
I think maybe it takes significantly longer to reencode the 3.5 second video that is recorded in the tests when running on CI/CD.  I hate having such a long wait it (now > 10s runtime) for the test, but.... This test doesn't run in dev anyway. It can and does run on linux, either in docker ubuntu.  I've been uploading it to a Pi5 with Debian Bookworm and running it there - the test finishes on Bookwork without error..

Testing the encoding duration theory by increasing the wait time to video duration * 3