make webRTC camera stream actually use the basic_bot.commons.base_camera frames
🎉 it works!!  the video streaming looks great in the test app.

still need to either make the old MJPEG video streaming work with the new aiohttp web server that replaced flask or maybe kill the old way.  The only bummer there is that it is dir simple to drop and <img tag on a page to use the old way - no js required.  I think maybe compromise and make it work with aiohttp but mark it as deprecated