test_vision integration tests should be able to run on dev (macOS)
the whole test module was previously being excluded ffrom pytest running on any computer that tflite runtime is not, can not be installed on like macOS dev env.  only one of the three tests require the recognition provider and tflight, instead of excluding whole module, only exclude the test that tests recognition.  This would have enabled me to see two of the failures before CI/CD