more robust vision service shutdown++
I noticed that the vision service was hanging on shutdown when running in the context of daphbot due while a video was recording.  I added more robust handling of stop() in base camera.  To determine that the base camera was the cause of the hang, I added a signal handler for USR1 that will show all of the threads and stack traces. I'm leaving it in this time.  If you read this commit message or stumble over it in the code, congratulations. To use: kill -USR1 <vision service pid>