failed attempt to use thread thread executer to serve MJPEG
Multiple MJPEG streams cause the WebRTC audio streams to start and stop and lag > 7 seconds.  My best analysis is that this is a problem because
MJPEG (multiple) and the WebRTC are sharing the same asyncio loop.  The MJPEG streams each start an async while true task that sleeps for just 1/30 second per frame.

I'm considering just using flask to serve up MJPEG requests from another thread.  I think a -2 branch is in the future.